requirements.txt
numpy>=1.0.0
matplotlib>=3.0.0
pandas>=1.2.0
PyYAML>=5.0.0
tabulate>0.8.0
